Split download into new file/module
`main.rs` was getting too big.
